2-python-microservices/requirements.txt
flask
flask-cors
openai-whisper
torch
torchaudio
transformers
edge-tts
pydub
audioop-lts
Pillow
opencv-python
opencv-contrib-python
face-alignment
librosa
scipy
numpy
numba
resampy
soundfile
tiktoken
imageio
imageio-ffmpeg
regex
